fix: refactor docstrings to abstract methods
